integrations/agentic-frameworks/crewai/pyproject.toml
[project]
name = "crewai-agent-example"
version = "1.0.0"
description = "CrewAI 多 Agent 协作系统集成 PPIO Agent Runtime 示例项目"
authors = [
    {name = "PPIO", email = "support@ppio.cloud"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "ppio-sandbox>=1.1.0b1",
    "crewai>=0.86.0",
    "crewai-tools>=0.15.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

